+++ packages/profile/src/components/context/index.tsx
@@ -1,3 +1,2 @@
-export { DataContext } from "./data";
 export { ThemeContext } from "./theme";
 export { Provider } from "./provider";
