+++ packages/profile/src/hooks/context.ts
@@ -1,6 +1,7 @@
 import { useContext } from "react";
-import { ThemeContext, DataContext } from "@/components/context";
+import { ThemeContext } from "@/components/context";
 import { ConnectionContext } from "@/context/connection";
+import { DataContext } from "@/context/data";
 
 export function useTheme() {
   return useContext(ThemeContext);
